+++ NEWS.md
@@ -7,3 +7,8 @@
 * Modified the `normalize_cell_counts` and `combine_cell_counts` function to also store counts and areas/volume by slice in the mouse and experiment objects, respectively. This allows for easy evaluation of the raw mapped dataset.
 * Our various analysis and visualization function are now compatible with the  Kim lab's [unified atlas](https://kimlab.io/brain-map/atlas/). The [unified](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6838086/) atlas synthesizes label from two of the most highly used atlases --Franklin-Paxinos (FP) and the common coordinate framework (CCF) from the Allen Institute.Check out the various help pages for analysis and plotting functions to see which ones now include`ontology` as a relevant parameter. Simply change the value from `allen` to `unified` to process datasets using this ontology.
 * New function `simplify_cell_count` can process external datasets and simplify regions based on keywords. Simplification is based on user indicated atlas (`allen` or `unified`)
+
+# SMARTR 1.0.1
+* New function `export_permutation_results` reformats permutation results and saves it nicely for users.
+* Tables are now saved into an autocreated `tables` folder in the output directory for organization.
+* Added `create_joined_networks` and `plot_joined_networks` as as capability to visualize the overlapping edges in two networks with outer joined nodes. The capability to export a list of overlapping edges between the joined network is included as a parameter in `create_joined_networks`. See the documentation for more details.